+++ src/lib.rs
@@ -1,24 +1,20 @@
-pub mod cli;
 pub mod commands;
 pub mod config;
 mod database;
 mod fix;
 mod fixers;
-mod issue;
-mod message;
-mod outcome;
+pub mod input;
+mod output;
 mod probes;
 pub mod testhelpers;
 
+pub use commands::{Issue, Outcome};
 use database::Tikibase;
-pub use fix::Fix;
-pub use issue::Issue;
-pub use message::{Message, Messages};
-pub use outcome::Outcome;
+pub use output::{Message, Messages};
 use std::path::PathBuf;
 
 /// runs the given Command in the given directory, returns structured data
-pub fn run(command: cli::Command, dir: PathBuf) -> Messages {
+pub fn run(command: input::Command, dir: PathBuf) -> Messages {
     let config = match config::load(&dir) {
         Ok(config) => config,
         Err(issue) => return Messages::from_issue(issue),
@@ -28,10 +24,10 @@ pub fn run(command: cli::Command, dir: PathBuf) -> Messages {
         Err(issues) => return Messages::from_issues(issues),
     };
     let outcome = match command {
-        cli::Command::Check => commands::check(&mut base, &config),
-        cli::Command::Stats => commands::stats(&base),
-        cli::Command::Fix => commands::fix(&mut base, &config),
-        cli::Command::Pitstop => commands::pitstop(&mut base, &config),
+        input::Command::Check => commands::check(&mut base, &config),
+        input::Command::Stats => commands::stats(&base),
+        input::Command::Fix => commands::fix(&mut base, &config),
+        input::Command::Pitstop => commands::pitstop(&mut base, &config),
     };
-    outcome.to_messages()
+    Messages::from_outcome(outcome)
 }
